Moved the Builders file into a package prior to breaking it into its sub-parts
