+++ kakao/src/main/kotlin/com/agoda/kakao/Assertions.kt
@@ -20,6 +20,8 @@ import android.support.test.espresso.web.model.Atom
 import android.support.test.espresso.web.model.ElementReference
 import android.support.test.espresso.web.sugar.Web
 import android.support.test.espresso.web.webdriver.DriverAtoms
+import com.agoda.kakao.builders.RootBuilder
+import com.agoda.kakao.builders.ViewBuilder
 import org.hamcrest.CoreMatchers
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers
